Game initilization now creates and places appropriate game pieces on the board. Also, a standard grid labal has been applied to the outside of the gamebaord.
Current issue - The "white" pieces appear to be clear with a white outline, while the "black" pieces
appear to be solid white. I have switched their positions on the board so that it appears the white
pieces are on the bottom. I need to switch them back, and then change the colors elsewhere so there
is no confusion about who is player 1 and 2 in the code.