Added a few things working towards en passant. There is a decay variable that if greater than 0, decays at 0.5 each turn. The goal is that en passant will not be a valid move if it's at 0.
There is also a functional method that runs at the start of each player's turn
that sets all of the player's pawns' vulnerable setting to false. This may make
the decay variable seem redundant, but having the variable will prevent running
all of the checks involved with allowing an en passant move.