Corrected an issue where the player's pieces array was not deleting the old king location after castling. This was causing faults where the AI would check for possible moves from a piece that didn't exist.
updated the in_check? method to use the get_enemy_moves method rather than try
to do it on it's own, simplifying code.

updated multiple instances where code was repeated depending on which player was
using it by assinging a couple specific variables values based on the player, and
just using the main code for both players, substituting in those variables. This has
cleaned up a good bit of code and there is likely more to go.

added a check in the AI to see if it was in check when it had no available moves. If
so, the game ends and "Checkmate!" is shown under the game board.